requirements.txt
imgui==2.0.0
numpy==2.3.5
pillow==12.0.0
pygame==2.6.1
pyglm==2.8.3
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
PyWavefront==1.3.3
trimesh==4.10.1
